Now with comments!
trying to fix issue where inserts are crashing the connection to database.  Latest version does this by chunking out the INSERTs limiting them to inserting 50 rows into a table at any time.  This is separate from the "Batch Size" variable that is now available on the command line.  The "Batch Size" variable limits the amount of rows out of the CSV the script parses at a time.