Beginings of http check ins.
Poller now lives in it's own folder (simulates being on a client machine) and has a config file. There is also a local template.xml file saved for the client.

A hash of the xml is compared to the hash of the template int he DB with the same php_name os, release and version of the client. If the hashes don't match, it grabs the template from the DB (all done via http calls to checkin.php).

Started inv_functions.php to store inventroy specific functions.  The first of which is getHashID.  You feed it a hash, and it checks to see if exists.  If it does, it gives you back the ID, if it doesn't, it inserts the hash into the database, then gives you the ID of the newly inserted line.

upadated template table

started list of hashes (DB)

-- to do
need to have the config autoupdate similar to the template.

poller neds to save local files to retry if dumping grounds aren't available.Beginings of http check ins.

Poller now lives in it's own folder (simulates being on a client machine) and has a config file. There is also a local template.xml file saved for the client.

A hash of the xml is compared to the hash of the template int he DB with the same php_name os, release and version of the client. If the hashes don't match, it grabs the template from the DB (all done via http calls to checkin.php).Beginings of http check ins.

Poller now lives in it's own folder (simulates being on a client machine) and has a config file. There is also a local template.xml file saved for the client.

A hash of the xml is compared to the hash of the template int he DB with the same php_name os, release and version of the client. If the hashes don't match, it grabs the template from the DB (all done via http calls to checkin.php).

Started inv_functions.php to store inventroy specific functions.  The first of which is getHashID.  You feed it a hash, and it checks to see if exists.  If it does, it gives you back the ID, if it doesn't, it inserts the hash into the database, then gives you the ID of the newly inserted line.

Updated template table in db, started the list of hashes.

-- to do
need to have the config autoupdate similar to the template.
set up checkin to accept result xml for dumping to db
make the poller write local files if it can't reach http/file dumping grounds and try to submit all the next time it is run.